mengimplementasikan jwt pada fitur delete product

+++ controllers/product_controllers.go
@@ -1,8 +1,10 @@
 package controllers
 
 import (
+	"log"
 	"net/http"
 	"project_altabe4_1/lib/databases"
+	"project_altabe4_1/middlewares"
 	"strconv"
 
 	"github.com/labstack/echo/v4"
@@ -40,14 +42,25 @@ func DeleteProductControllers(c echo.Context) error {
 			"message": "False Param",
 		})
 	}
-
+	id_user_product, _ := databases.GetIDUserProduct(conv_id)
+	log.Println("id_user_product", id_user_product)
+	logged := middlewares.ExtractTokenId(c)
+	log.Println("idlogged", logged)
+	if uint(logged) != id_user_product {
+		return c.JSON(http.StatusBadRequest, map[string]interface{}{
+			"code":    http.StatusBadRequest,
+			"message": "Access Forbidden",
+		})
+	}
 	_, e := databases.DeleteProduct(conv_id)
+
 	if e != nil {
 		return c.JSON(http.StatusBadRequest, map[string]interface{}{
 			"code":    http.StatusBadRequest,
 			"message": "Bad Request",
 		})
 	}
+
 	return c.JSON(http.StatusOK, map[string]interface{}{
 		"code":    http.StatusOK,
 		"message": "Successful Operation",
